Added metaclass
