+++ test.json
@@ -1,1 +1,1 @@
-{"a": 1, "b": 1, "c": {"a": 1, "b": 2}}
+{"a": 1, "b": 1, "c": {"a": 1, "b": 2}, "bc": {"g": 5}}
